moved all of the docker stack code out of lores-node-axum
