+++ backend/coop-cloud-docker-apps/src/docker/docker_stack.rs
@@ -1,6 +1,6 @@
 use std::process::Command;
 
-use super::{DockerStack, helpers::parse_docker_json};
+use super::{DockerService, DockerStack, helpers::parse_docker_json};
 
 #[derive(Debug, Clone, serde::Deserialize)]
 #[allow(dead_code)]
@@ -68,4 +68,74 @@ pub fn docker_stack_services(
     let services = parse_docker_json::<Vec<DockerStackServicesResult>>(output)?;
 
     Ok(services)
+}
+
+#[derive(Debug, Clone, serde::Deserialize)]
+#[allow(dead_code)]
+pub struct DockerStackPsResult {
+    #[serde(rename = "ID")]
+    pub id: String,
+
+    #[serde(rename = "Name")]
+    pub name: String,
+
+    #[serde(rename = "Image")]
+    pub image: String,
+
+    #[serde(rename = "Node")]
+    pub node: String,
+
+    #[serde(rename = "DesiredState")]
+    pub desired_state: String,
+
+    #[serde(rename = "CurrentState")]
+    pub current_state: String,
+
+    #[serde(rename = "Error")]
+    pub error: Option<String>,
+
+    #[serde(rename = "Ports")]
+    pub ports: String,
+}
+
+pub fn docker_stack_ps(stack_name: &str) -> Result<Vec<DockerService>, anyhow::Error> {
+    let output = Command::new("docker")
+        .arg("stack")
+        .arg("ps")
+        .arg(stack_name)
+        .arg("--format")
+        .arg("json")
+        .arg("--filter")
+        .arg("desired-state=Running")
+        .output()
+        .map_err(|e| anyhow::anyhow!("Failed to execute command: {}", e))?;
+
+    let services = parse_docker_json::<Vec<DockerStackPsResult>>(output)?;
+
+    let services: Vec<DockerService> = services
+        .into_iter()
+        .map(|result| {
+            let (current_state, current_state_duration) =
+                split_state_and_duration(&result.current_state);
+            DockerService {
+                id: result.id,
+                name: result.name,
+                image: result.image,
+                node_name: result.node,
+                current_state,
+                current_state_duration,
+            }
+        })
+        .collect();
+
+    Ok(services)
+}
+
+fn split_state_and_duration(state: &str) -> (String, String) {
+    let parts: Vec<&str> = state.splitn(2, ' ').collect();
+    if parts.len() == 2 {
+        (parts[0].to_string(), parts[1].to_string())
+    } else {
+        (state.to_string(), String::new())
+    }
 }
